General: Use forward declarations where applicable
Also removes header inclusions where they aren't necessary, of which a
few have cropped up just as a matter of the project evolving over time.

This minimizes the amount of files that need to be rebuilt when changes
to the implementation occur.

No behavioral changes are made, this is purely code movement.