Features for the Ammo Drops and Heart Drops and Refills
With the Ammo Drops option set to Off, ammo drops are turned into blue rupees, except for Deku Stick because of the special Deku Baba drops. This does not affect getItem items.

With the Heart Drops and Refills option set to Off, hearts are turned into green rupees. This does not affect getItem Recovery Hearts.

To keep the randomiser application UI condensed, I integrated the Bombchu Drops option with the Ammo Drops option as a possible setting (On + Bombchu).