Adapted the logic for Ammo Drops
With Ammo Drops set to Off, slingshot, bow, bombs, bombchu and deku nuts require ammo refills to be available in a shop.
Magic items require a bottle and Green Potion to be available in a shop. While magic can be replenished in other locations, some instances (such as the Ganondorf fight) may require more magic than a single magic gauge can provide, so this is a failsafe.

Bombchu in logic no longer requires bombchu drops, since the logic ensures that bombchu refills can be found in shops if drops are unavailable.