Added Shopsanity Prices setting
Setting is either Random: The default and original shopsanity behaviour; or Affordable: all shopsanity items are priced at 10 ruppees, akin to affordable scrub sanity.

I would like to disable/hide the setting if the shopsanity setting is not set to an option that enables the shopsanity items, but for now I don't know how yet.