Added Shopsanity Prices setting (#774)
* Added Shopsanity Prices setting

Setting is either Random: The default and original shopsanity behaviour; or Affordable: all shopsanity items are priced at 10 ruppees, akin to affordable scrub sanity.

I would like to disable/hide the setting if the shopsanity setting is not set to an option that enables the shopsanity items, but for now I don't know how yet.

* Added hiding/unhiding of Shopsanity prices option

Shopsanity prices option is now hidden when Shopsanity is set to either off or zero

* Update source/descriptions.cpp

Co-authored-by: Colin E. <5352197+Kewlan@users.noreply.github.com>

* Update source/settings.cpp

Co-authored-by: Colin E. <5352197+Kewlan@users.noreply.github.com>

* Moved show/hide of Shopsanity prices setting up

* Run automatic format script as code does not match clang format rules.

Developers please rebase to avoid merge conflicts!

* Update source/settings.cpp

---------

Co-authored-by: Colin E. <5352197+Kewlan@users.noreply.github.com>
Co-authored-by: viccie211 <viccie211@users.noreply.github.com>