Feature/shop sanity wallet price caps (#776)
* Added expensive option for shopsanity

Default random behaviour generates an int between 0 and 59 in a distribution and then multiplies that by 5 to get prices between 0 and 295 in increments of 5. This expensive option changes that multiplier to 10 so it ranges between 0 and 590 in increments of 10 rupees

* Added Shopsanity price cap setting for each wallet

* Reworded descriptions to be the same accross the options

* Run automatic format script as code does not match clang format rules.

Developers please rebase to avoid merge conflicts!

* Moved affordable into switch and added comments

The switch now handles all cases, affordable just short circuits and returns 10. Technically this is faster since it doesn't need to always call .Is() for the affordable option even in the cases that it isn't set to affordable

* Run automatic format script as code does not match clang format rules.

Developers please rebase to avoid merge conflicts!

* Restored original behaviour and added all possible values.

For the wallet cap options I now run the calculation of the probability function on the fly instead of using the precomputed table. Which gives the possibility for all possible 5 rupee increments to happen in each chosen setting, at the cost of some compute time.

* Restored original behaviour and added all possible values.

For the wallet cap options I now run the calculation of the probability function on the fly instead of using the precomputed table. Which gives the possibility for all possible 5 rupee increments to happen in each chosen setting, at the cost of some compute time.

* Formatting suggestion

Co-authored-by: HylianFreddy <82058772+HylianFreddy@users.noreply.github.com>

* Formatting suggestion

Co-authored-by: HylianFreddy <82058772+HylianFreddy@users.noreply.github.com>

* Formatting suggestion

Co-authored-by: HylianFreddy <82058772+HylianFreddy@users.noreply.github.com>

---------

Co-authored-by: viccie211 <viccie211@users.noreply.github.com>
Co-authored-by: HylianFreddy <82058772+HylianFreddy@users.noreply.github.com>